requirements.txt
streamlit
pandas
matplotlib
seaborn
google-generativeai
openpyxl